Melee Enemy Implemented
implemented attack for melee enemies, can be adjusted for other types of enemies
