+++ Assets/Scripts/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d64a5b2ee4a538439c4d42a18ad0854
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
