+++ Assets/Scripts/Enemies/MeleeEnemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b70bafe941dc984b89ccd347d17b28e
+guid: 19ba255a96d54ed439f20f7ba89f919d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
