--- Assets/Scripts/EnemyDamage.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyDamage : MonoBehaviour
-{
-    [SerializeField]
-    private int damage;
-    [SerializeField]
-    private PlayerHealth playerHealth;
-
-    private void OnCollisionEnter2D(Collision2D collision)
-    {
-        if(collision.gameObject.tag == "Player")
-        {
-            playerHealth.TakeDamage(damage);
-        }
-    }
-}
