+++ Assets/Scripts/Player/PlayerCombat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bdd1ce730c1560e42902adb89c7baf59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
